Edit: Moved sleectedGame entities to domain layer
