+++ src/modules/selectedGame/domain/entities/IActivity.ts
