+++ src/modules/selectedGame/domain/entities/IFeedPost.ts
@@ -1,5 +1,5 @@
-import IAchievement from 'modules/selectedGame/entities/IAchievement';
-import IActivity from 'modules/selectedGame/entities/IActivity';
+import IAchievement from 'modules/selectedGame/domain/entities/IAchievement';
+import IActivity from 'modules/selectedGame/domain/entities/IActivity';
 import IPlayer from 'shared/domain/entities/IPlayer';
 import IRank from 'shared/domain/entities/IRank';
 
@@ -12,7 +12,7 @@ interface IFeedPost {
   level: {
     level: number;
     requiredExperience: number;
-    title?: string
+    title?: string;
   };
   rank: IRank;
   date: string;
