+++ src/modules/selectedGame/infra/repositories/AchievementsRepository.ts
@@ -1,4 +1,4 @@
-import IAchievement from 'modules/selectedGame/entities/IAchievement';
+import IAchievement from 'modules/selectedGame/domain/entities/IAchievement';
 import IAchievementsRepository from 'modules/selectedGame/repositories/IAchievementsRepository';
 import makeHttpProvider from 'shared/infra/providers/factories/makeHttpProvider';
