+++ src/modules/selectedGame/repositories/IAchievementsRepository.ts
@@ -1,4 +1,4 @@
-import IAchievement from 'modules/selectedGame/entities/IAchievement';
+import IAchievement from 'modules/selectedGame/domain/entities/IAchievement';
 
 export default interface IAchievementsRepository {
   findAll(): Promise<IAchievement[]>;
