+++ src/modules/selectedGame/repositories/IFeedPostsRepository.ts
@@ -1,4 +1,4 @@
-import IFeedPost from 'modules/selectedGame/entities/IFeedPost';
+import IFeedPost from 'modules/selectedGame/domain/entities/IFeedPost';
 
 export default interface IFeedPostsRepository {
   findAll(): Promise<IFeedPost[]>;
