+++ src/modules/selectedGame/repositories/ILeaderboardsRepository.ts
@@ -1,4 +1,4 @@
-import ILeaderboard from 'modules/selectedGame/entities/ILeaderboard';
+import ILeaderboard from 'modules/selectedGame/domain/entities/ILeaderboard';
 
 export default interface ILeaderboardsRepository {
   findCurrent(): Promise<ILeaderboard | null>;
