+++ src/modules/selectedGame/view/pages/Activities/index.tsx
@@ -11,7 +11,7 @@ import useGetActivitiesController from 'modules/selectedGame/infra/controllers/u
 
 // Styles
 import { Container, PageTitle, Description, ActivityContainer } from './styles';
-import IActivity from 'modules/selectedGame/entities/IActivity';
+import IActivity from 'modules/selectedGame/domain/entities/IActivity';
 
 const Activities: React.FC = () => {
   const { loading, activities, getActivities } = useGetActivitiesController();
