+++ src/modules/selectedGame/view/pages/Feed/components/AchievementFeed/index.tsx
@@ -1,4 +1,4 @@
-import IFeedPost from 'modules/selectedGame/entities/IFeedPost';
+import IFeedPost from 'modules/selectedGame/domain/entities/IFeedPost';
 import React from 'react';
 
 // Components
