+++ src/modules/selectedGame/view/pages/Feed/components/ActivityFeed/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import IFeedPost from 'modules/selectedGame/entities/IFeedPost';
+import IFeedPost from 'modules/selectedGame/domain/entities/IFeedPost';
 import FeedText from '../FeedText';
 import RankTag from '../RankTag';
