+++ src/modules/selectedGame/view/pages/Ranking/components/RankingPosition/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { Container, Icon, Image, PositionBlock, Text } from './styles';
 
-import { IPosition } from 'modules/selectedGame/entities/ILeaderboard';
+import { IPosition } from 'modules/selectedGame/domain/entities/ILeaderboard';
 import { RankChip } from 'modules/selectedGame/view/components';
 import { getTextColor } from 'shared/view/helpers';
